+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added cached token information for Gemini models
 - fixed issue with unfinished stream response text preceding a tool call
 - default agent profile for new users set to Power Tools and added Aider as the second one
+- explicit cache control with Requesty Anthropic models used only when auto caching is disabled
 
 ## [0.20.0]